fixed compiler warnings
git-svn-id: svn+ssh://svn.pointclouds.org/pcl/trunk@7230 a9d63959-f2ad-4865-b262-bf0e56cfafb6

+++ registration/include/pcl/registration/impl/correspondence_estimation_backprojection.hpp
@@ -126,7 +126,7 @@ pcl::registration::CorrespondenceEstimationBackProjection<PointSource, PointTarg
         if (dist < min_dist)
         {
           min_dist = dist;
-          min_index = j;
+          min_index = static_cast<int> (j);
         }
       }
       if (min_dist > max_distance)
@@ -167,7 +167,7 @@ pcl::registration::CorrespondenceEstimationBackProjection<PointSource, PointTarg
         if (dist < min_dist)
         {
           min_dist = dist;
-          min_index = j;
+          min_index = static_cast<int> (j);
         }
       }
       if (min_dist > max_distance)
@@ -241,7 +241,7 @@ pcl::registration::CorrespondenceEstimationBackProjection<PointSource, PointTarg
         if (dist < min_dist)
         {
           min_dist = dist;
-          min_index = j;
+          min_index = static_cast<int> (j);
         }
       }
       if (min_dist > max_distance)
@@ -289,7 +289,7 @@ pcl::registration::CorrespondenceEstimationBackProjection<PointSource, PointTarg
         if (dist < min_dist)
         {
           min_dist = dist;
-          min_index = j;
+          min_index = static_cast<int> (j);
         }
       }
       if (min_dist > max_distance)
